Add ability to search for recipes that will pair well
Set up a node scraper to grab ingredient pairings. A random ingredient from the ingredients based on current kebab is grabbed, corresponding well-paired ingredients are fetched from scraper, and a random well-paired ingredient is used to fetch new recipes that will, in theory, pair well. It gets fun when you search for an ingredient that pairs well with flour. Or salt...